fix many bugs in install system

+++ install/index.php
@@ -12,7 +12,6 @@
 include important files
 */
 
-
 define ( 'IN_COMMON' , true);
 $_path = "../";
 (file_exists($_path . 'config.php')) ? include_once ($_path . 'config.php') : null;
@@ -49,54 +48,65 @@ switch ($_GET['step'])
 {
 default:
 case 'language':
-		if (isset($_POST['lang'])) 
+
+	if(isset($_POST['lang']))
+	{
+		if (!empty($_POST['lang']))
 		{
-			if (!empty($_POST['lang']))
-			{
 				//go to .. 2step
 				echo '<meta http-equiv="refresh" content="0;url=' . $_SERVER['PHP_SELF'] . '?step=choose&lang=' . $_POST['lang'] . '">';
-			}
 		}
-		else //no language
-		{ 
-
-	//get language from LANGUAGE folder
-	$path = "../lang";
-	$lngfiles = '';
-	if ($dh = @opendir($path))
-	{
-		while (($file = readdir($dh)) !== false)
-		{
-			if(strpos($file, '.') === false && $file != '..' && $file != '.')
-			{
-				$lngfiles .= '<option value="' . $file . '"' . ($file == 'en' ? ' selected="selected"' : '') . '>' . $file . '</option>';
-			}
-		}
-		
-		@closedir($dh);
+	
 	}
+	//no language begin
+	else
+	{
+		//get language from LANGUAGE folder
+		$path = "../lang";
+		$lngfiles = '';
+		if ($dh = @opendir($path))
+		{
+			while (($file = readdir($dh)) !== false)
+			{
+				if(strpos($file, '.') === false && $file != '..' && $file != '.')
+				{
+					$lngfiles .= '<option value="' . $file . '"' . ($file == 'en' ? ' selected="selected"' : '') . '>' . $file . '</option>';
+				}
+			}
+			
+			@closedir($dh);
+		}
 
 
-	// show  language list ..
-	echo '<br />		
-	<div class="centery">
-		<fieldset class="home">
-			<img src="img/map.png" style="border:0" alt="al-Idrisi Map">
-			<br />
-			<form  action="' . $_SERVER['PHP_SELF'] . '?step=language&' . getlang(1) . '" method="post">
-				<select name="lang" style="width: 352px">' . $lngfiles . '</select>
-				<br /><br /><br /><input name="submitlang" type="submit" value=" [  Next >>  ] " /><br /><br /><br />
-			</form>
-		</fieldset>
-	</div>';
+		// show  language list ..
+		echo '<br />		
+		<div class="centery">
+			<fieldset class="home">
+				<img src="img/map.png" style="border:0" alt="al-Idrisi Map">
+				<br />
+				<form  action="' . $_SERVER['PHP_SELF'] . '?step=language&' . getlang(1) . '" method="post">
+					<select name="lang" style="width: 352px">' . $lngfiles . '</select>
+					<br /><br /><br /><input name="submitlang" type="submit" value=" [  Next >>  ] " /><br /><br /><br />
+				</form>
+			</fieldset>
+		</div>';
 
-		}//no language else
+	}
+	//no language end
 
 
 
-break; // end case language
+break;
 case 'choose' :
-
+		
+		
+	//check version of PHP 
+	if (!function_exists('version_compare') || version_compare(PHP_VERSION, MIN_PHP_VERSION, '<'))
+	{
+		echo '<fieldset class="home"><span style="color:red;">' . sprintf($lang['INST_PHP_LESSMIN'], MIN_PHP_VERSION, PHP_VERSION) . '</span><br /></fieldset>';
+	}
+	else
+	{
 		echo '<fieldset class="home"><span style="color:green;">' . $lang['INST_CHOOSE_INSTALLER'] . '</span><br /><br /><br />';
 		
 		$install_or_no = true;
@@ -122,7 +132,7 @@ case 'choose' :
 		echo '<a href="./update.php?' . getlang(1) . '"><img src="img/updater.png" alt="updater" /> <br /> ' . $lang['INST_UPDATE_P_VER'] . ' </a><br /><br /><br />';
 
 		echo '<a href="http://www.kleeja.com"><span style="color:black;">www.kleeja.com</span></a></fieldset>';
-
+	}
 break;
 
 }#endOFswitch
@@ -134,5 +144,3 @@ break;
 */
 echo $footer_inst;
 
-
-?>
